+++ gigaspatial/handlers/giga/__init__.py
@@ -1,9 +1,11 @@
 from gigaspatial.handlers.giga.school_locations import GigaSchoolLocationFetcher
 from gigaspatial.handlers.giga.school_measurements import GigaSchoolMeasurementsFetcher
 from gigaspatial.handlers.giga.school_profile import GigaSchoolProfileFetcher
+from gigaspatial.handlers.giga.countries import GigaCountriesFetcher
 
 __all__ = [
     "GigaSchoolLocationFetcher",
     "GigaSchoolMeasurementsFetcher",
     "GigaSchoolProfileFetcher",
+    "GigaCountriesFetcher",
 ]